requirements.txt
django
django_rest_framework
django-rest-swagger
django_filter
markdown
pygments
